+++ src/main/java/romever/scan/oasisscan/repository/AccountRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Integer>,
+public interface AccountRepository extends JpaRepository<Account, Long>,
         JpaSpecificationExecutor<Account> {
 
     Optional<Account> findByAddress(String address);
